pretty output

+++ main.py
@@ -23,25 +23,45 @@ def main():
     canvas = Canvas(api_url, api_key)
     course = canvas.get_course(input("Enter course id: "))
 
-    #getting all files
-    print("fetching files")
+    print("Fetching files")
     files = []
     for file in course.get_files():
             files.append(File(file))
 
-    print("finding duplicates")
+    print("Finding duplicates")
     dups = []
+    counts = []
     for file in files:
         if files.count(file) > 1 and file not in dups:
             dups.append(file)
+            counts.append(files.count(file))
 
-    print("deleting duplicates")
-    for file in files:
-        if file in dups:
-            dup_id = dups.index(file)
-            if file.id != dups[dup_id].id:
-                print(f"deleting {file}")
-                file.delete()            
+    if dups == []:
+        print("No duplicates found")
+        quit()
+    else:
+        print("Duplicates found")
+
+        for i in range(len(dups)):
+            print("")
+            print("{:<50} {:>50}".format("FILENAME", "COUNT"))
+            print("{:<50} {:>50}".format("--------", "-----"))
+            print("{:<50} {:>50}".format(dups[i].filename, counts[i]))
+            print("")
+
+        prompt = input("Duplicate files will now be deleted. Continue?[y/N]: ")
+
+        if prompt == "y":
+            print("Deleting duplicates")
+            for file in files:
+                if file in dups:
+                    dup_id = dups.index(file)
+                    if file.id != dups[dup_id].id:
+                        print(f"deleting {file}")
+                        file.delete()
+        else:
+            print("abort.")
+            quit()            
             
     
 if __name__ == "__main__":
